Refactor chart files
